+++ lib/lifx_http_api.dart
@@ -2,3 +2,4 @@ library lifx_http_api;
 
 export './src/bulb.dart';
 export './src/client.dart';
+export './src/color.dart';