+++ lib/src/color.g.dart
@@ -6,15 +6,15 @@ part of 'color.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Color _$ColorFromJson(Map<String, dynamic> json) {
-  return Color(
+LifxColor _$LifxColorFromJson(Map<String, dynamic> json) {
+  return LifxColor(
     hue: (json['hue'] as num).toDouble(),
     saturation: (json['saturation'] as num).toDouble(),
     kelvin: json['kelvin'] as int,
   );
 }
 
-Map<String, dynamic> _$ColorToJson(Color instance) => <String, dynamic>{
+Map<String, dynamic> _$LifxColorToJson(LifxColor instance) => <String, dynamic>{
       'hue': instance.hue,
       'saturation': instance.saturation,
       'kelvin': instance.kelvin,